requirements.txt
yt-dlp>=2023.0.0
ffmpeg-python>=0.2.0
